refactor: split DesktopSecondaryMenuSlot into v1/v2
