+++ src/plugin-slots/DesktopSecondaryMenuSlot/index.jsx
@@ -1,40 +1,4 @@
-import React from 'react';
-import { PluginSlot } from '@openedx/frontend-plugin-framework';
-import DesktopHeaderMainOrSecondaryMenu, { desktopHeaderMainOrSecondaryMenuDataShape } from '../../desktop-header/DesktopHeaderMainOrSecondaryMenu';
-import HeaderNotificationsSlot from '../HeaderNotificationsSlot';
+import DesktopSecondaryMenuSlot from './v2';
 
-export const DesktopSecondaryMenuSlotV1 = ({
-  menu,
-}) => (
-  <PluginSlot
-    id="org.openedx.frontend.layout.header_desktop_secondary_menu.v1"
-    idAliases={['desktop_secondary_menu_slot']}
-    slotOptions={{
-      mergeProps: true,
-    }}
-  >
-    <DesktopHeaderMainOrSecondaryMenu menu={menu} />
-  </PluginSlot>
-);
-
-DesktopSecondaryMenuSlotV1.propTypes = {
-  menu: desktopHeaderMainOrSecondaryMenuDataShape,
-};
-
-const DesktopSecondaryMenuSlot = ({
-  menu,
-}) => (
-  <PluginSlot
-    id="org.openedx.frontend.layout.header_desktop_secondary_menu.v2"
-    pluginProps={{ menu }}
-  >
-    <HeaderNotificationsSlot />
-    <DesktopSecondaryMenuSlotV1 menu={menu} />
-  </PluginSlot>
-);
-
-DesktopSecondaryMenuSlot.propTypes = {
-  menu: desktopHeaderMainOrSecondaryMenuDataShape,
-};
-
+export { default as DesktopSecondaryMenuSlotV1 } from './v1';
 export default DesktopSecondaryMenuSlot;
