+++ src/plugin-slots/HeaderNotificationsSlot/README.md
@@ -12,10 +12,10 @@ This slot renders the notifications tray (bell icon + notification popover) from
 1. **Desktop Header** — via `org.openedx.frontend.layout.header_desktop_secondary_menu.v2`
    Notifications appear before secondary menu items (e.g., "New", "Help")
 
-2. **Learning Header** — via `org.openedx.frontend.layout.learning_header_actions.v2`
+2. **Learning Header** — via `org.openedx.frontend.layout.learning_header_actions.v1`
    Notifications appear before the help link
 
-3. **Studio Header** — via `org.openedx.frontend.layout.studio_header_actions.v2`
+3. **Studio Header** — via `org.openedx.frontend.layout.studio_header_actions.v1`
    Notifications appear before the search button
 
 ## Slot Hierarchy
@@ -27,14 +27,14 @@ Desktop Header
     └── org.openedx.frontend.layout.header_desktop_secondary_menu.v1 (menu items only)
 
 Learning Header
-└── org.openedx.frontend.layout.learning_header_actions.v2
+└── org.openedx.frontend.layout.learning_header_actions.v1
     ├── org.openedx.frontend.layout.header_notifications_tray.v1 ← This slot
-    └── org.openedx.frontend.layout.learning_header_actions.v1 (help link only)
+    └── org.openedx.frontend.layout.header_learning_help.v1
 
 Studio Header
-└── org.openedx.frontend.layout.studio_header_actions.v2
+└── org.openedx.frontend.layout.studio_header_actions.v1
     ├── org.openedx.frontend.layout.header_notifications_tray.v1 ← This slot
-    └── org.openedx.frontend.layout.studio_header_actions.v1 (search button only)
+    └── org.openedx.frontend.layout.studio_header_search_button_slot.v1
 ```
 
 ## Examples
