+++ src/plugin-slots/StudioHeaderActionsSlot/index.tsx
@@ -7,25 +7,14 @@ interface StudioHeaderActionsSlotProps {
   searchButtonAction?: React.MouseEventHandler<HTMLButtonElement>;
 }
 
-export const StudioHeaderActionsSlotV1: FunctionComponent<StudioHeaderActionsSlotProps> = ({
+const StudioHeaderActionsSlot: FunctionComponent<StudioHeaderActionsSlotProps> = ({
   searchButtonAction,
 }) => (
   <PluginSlot
     id="org.openedx.frontend.layout.studio_header_actions.v1"
   >
-    <StudioHeaderSearchButtonSlot searchButtonAction={searchButtonAction} />
-  </PluginSlot>
-);
-
-const StudioHeaderActionsSlot: FunctionComponent<StudioHeaderActionsSlotProps> = ({
-  searchButtonAction,
-}) => (
-  <PluginSlot
-    id="org.openedx.frontend.layout.studio_header_actions.v2"
-    pluginProps={{ searchButtonAction }}
-  >
     <HeaderNotificationsSlot />
-    <StudioHeaderActionsSlotV1 searchButtonAction={searchButtonAction} />
+    <StudioHeaderSearchButtonSlot searchButtonAction={searchButtonAction} />
   </PluginSlot>
 );
